Merging front&back

+++ NeovaDashboard/Source/UI/SubComponents/Gesture/MidiPanel.h
@@ -94,7 +94,7 @@ private:
 
     DraggableObject objectBeingDragged = none;
     Colour highlightColour;
-	NormalisableRange<int> midiRange = {0, 127};
+	NormalisableRange<float> midiRange = {0.f, 127.f};
 
     //==============================================================================
     JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR (MidiRangeTuner)
